Native (build): Simplify the build procedure
New cinterop options were added for us to use. It allows to compile
arbitrary files using the same flags to clang++ as those that are
used to build the Kotlin/Native programs themselves. Without
recreating those flags, linking is usually impossible, and so
before we had to rebuild everything manually.

For details on the new cinterop flags, see
https://github.com/JetBrains/kotlin-native/commit/0deca04ce7ad6e4d48c00d4c1c7545cbdfab1214#diff-baf040396c9558f79b08b46cc838f047