Native (Windows): provide an initial implementation
Common defines had to be moved to a separate file because on
Windows `date` does not build with `USE_OS_TZDB`. Even though the
timezone library is not used, it is unclear how to tell `cinterop`
not to build it for Windows, so now it is made to build by
unsetting `USE_OS_TZDB`.

A new gradle task was introduced for generating the
`windows_zones.hpp` file using the information from the Unicode
repository. This file contains the mapping between the names of
timezones used by Windows and the standard IANA timezone names.

A test had to be fixed since the previous version of the test
failed on Windows even though the code was correct because Windows
doesn't have accurate information about the `Asia/Gaza` timezone.