Native (Windows): cache Windows timezone info
This uses a read-write lock to guard the cache because
`unordered_map` is not thread-safe.

Another option would be to use TLS (thread-local storage), but the
clang bundled with Kotlin/Native does not support complex
destructors for objects in TLS.

Using the read-write lock required to use C++17, which led to
another problem: the C++ standard library bundled with
Kotlin/Native is too old on some platforms to have all the
definitions needed for C++17, and so the `date` library,
recognizing that the features of C++17 are available and attempting
to use them, failed to build. To fix this, a dirty hack was
introduced: on Unix platforms, the `__cplusplus` macro was
redefined to indicate that C++11 is used. This behaviour is
widely discouraged, but given that the version of the compiler used
is controlled by us and anomalies would bite us before users even
know about them, this seems like a valid tradeoff.