Use nanosecond resolution for Instant
Before, Instant did use nanosecond precision internally, and it
could be accessed by adding or substracting time periods or
constructing a new Instant by parsing a string with a nanosecond
component. Now, all the internal state of an `Instant` is more
easily accessible.

Also, for consistency, '*UnixMillis' was renamed to
'*EpochMilliseconds' for consistency: it is customary to use
the full measurement unit names in Kotlin libraries. Additionally,
what was called converters to and from Unix time was actually
using count of millisecond since 01 Jan 1970 in our epoch
(UTC-SLS), so that part of the name was changed accordingly.