Native (Apple): Provide integrations with the native datetime (#10)
For now, these are provided as a separate artifact. This is cumbersome
for the end user, but currently, since there is no ability yet to specify
common native code, we would have to duplicate much between every
available platform just to provide some special behavior on
Darwin-based devices. There is a clear migration path once common
native code becomes possible, so the problem is not significant.