Encode the form of formats for parsing and formatting anything
`FormatStructure` defines how a thing is parsed and/or formatted.
It encodes a group of `FieldFormatDirective` instances and directs
how they should behave as a collective:
* should they be parsed/output optionally?
* can there be several alternatives during parsing?
* etc.

The intended way to use the format machinery is to specify a
`FormatStructure` instance and obtain a parser or a formatter by
calling `parser()` or `formatter()` on it.

`CachedFormatStructure` can be used as a thin caching wrapper,
ensuring `parser()` and `formatter()` don't get called repeatedly.